Move volgen to bin/glusterd2/volgen
