+++ bin/glusterd2/volgen/static.go
