+++ bin/glusterd2/volgen/varstrings.go
