+++ commands/volumes/common.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
+	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
 	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/servers/sunrpc"
-	"github.com/gluster/glusterd2/volgen"
 	"github.com/gluster/glusterd2/volume"
 	"github.com/gluster/glusterd2/xlator"
